+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditor.java
@@ -300,9 +300,8 @@ public class EntryEditor extends JPanel implements EntryContainer {
         otherFields.remove(BibEntry.KEY_FIELD);
         otherFields.removeAll(Globals.prefs.getCustomTabFieldNames());
 
-        if (!otherFields.isEmpty()) {
-            addOtherTab(otherFields);
-        }
+        otherFields.add("otherfield");
+        addOtherTab(otherFields);
 
         // general fields from preferences
         addGeneralTabs();